Moved TransientUI to transient_sys_id_ui.py
